Create a Custom Ordered Variable

This function transforms a numeric variable into an ordered factor variable based on specified breaks and labels. It is useful for categorizing a continuous variable into a discrete, ordered factor variable. The function adds the newly created ordered factor variable to the dataframe with a modified name indicating the transformation.

Usage

create_ordered_variable(df, var_name, breaks, labels)

Arguments

df

A dataframe containing the variable to be transformed.

var_name

The name of the numeric variable in `df` to be transformed into an ordered factor.

breaks

A numeric vector of values specifying the breakpoints between categories. The length of `breaks` should be one more than the length of `labels`.

labels

A character vector specifying the labels for the resulting categories. These labels correspond to intervals defined by `breaks`.

Value

The input dataframe with an added column for the newly created ordered factor variable. This column is named by appending "_coarsen" to the original variable name.

Examples

# Assuming `dt` is a dataframe with a numeric variable "hours_work"
dt <- data.frame(hours_work = c(20, 35, 10, 45))
dt <- create_ordered_variable(dt, "hours_work", c(10, 30, 41, Inf), c("[10_30)", "[30_41)", "[41_up]"))
print(dt)
#>   hours_work hours_work_coarsen
#> 1         20            [10_30)
#> 2         35            [30_41)
#> 3         10            [10_30)
#> 4         45            [41_up]