Create Combined Tables for Subset Condition Using GRF Models

This function calculates new evaluation tables for specified outcome variables using a subset condition, and combines them into a single data frame. It uses the models generated by `model_grf_causal_forest` or similar functions to create these tables.

Usage

margot_grf_subset_table(
  model_results,
  outcome_vars = NULL,
  subset_condition = NULL,
  scale = "RD",
  debug = FALSE
)

Arguments

model_results

A list of model results from `model_causal_forest` or similar functions. Note that these functions should be called with `save_models = TRUE`.

outcome_vars

Optional. A character vector of outcome variable names. If NULL, the function will use all models in the input.

subset_condition

A logical vector indicating the subset of data to use. Default is NULL.

scale

A character string indicating the scale to use in `margot_model_evalue`. Default is "RD".

debug

Logical. If TRUE, print debug information. Default is FALSE.

Value

A data frame combining all custom evaluation tables for the specified subset condition. If no full models are available, returns NULL and prints a reminder to set save_models = TRUE.

Details

The function uses `margot_model_evalue` to calculate evaluation metrics for each model in the subset of data specified by `subset_condition`. If outcome_vars is not provided, it will use all models in the input. The function can handle models with or without the "model_" prefix.

Note

Ensure that the `model_results` list contains full model objects for all specified `outcome_vars`. This requires setting `save_models = TRUE` when calling the original GRF model function.

See also

Examples

if (FALSE) { # \dontrun{
# Assuming result_outcomes_social_test is your results object from model_grf_causal_forest
subset_condition <- X_toy[, "t0_self_esteem_z"] < 0
result_subset <- margot_grf_subset_table(
  model_results = result_outcomes_social_test,
  outcome_vars = outcomes_social,
  subset_condition = subset_condition,
  scale = "RD"
)
print(result_subset)
} # }