Create Plot Options for Margot Plot

This function creates a list of options for use with the `margot_plot` function. It allows for easy customization of plot settings while maintaining a set of default values.

Usage

margot_plot_create_options(
  subtitle,
  base_defaults = NULL,
  title = NULL,
  filename_prefix = NULL,
  ...
)

Arguments

subtitle

Character string. The subtitle for the plot.

base_defaults

List. The base default options to use. If not provided, uses a pre-defined set of defaults.

title

Character string. The title for the plot. If NULL, uses the title from base_defaults.

filename_prefix

Character string. Prefix for the filename. If NULL, uses a default prefix.

...

Additional named arguments to override or add to the default options.

Value

A list of plot options to be used with `margot_plot`.

Examples

if (FALSE) { # \dontrun{
health_options <- margot_plot_create_options("Health Outcomes")
education_options <- margot_plot_create_options("Education Outcomes",
                                        title = "Custom Title",
                                        filename_prefix = "edu_outcomes")
} # }