Create a Combined Decision Tree and Policy Relationship Graph

This function generates a combined plot consisting of a decision tree and a graph showing relationships between variables in the recommended policy.

Usage

margot_plot_policy_combo(
  mc_test,
  model_name,
  layout = list(heights = c(1, 2)),
  annotation = list(tag_levels = "A")
)

Arguments

mc_test

An object containing Monte Carlo test results (exact type/class to be specified)

model_name

A character string specifying the name of the model

layout

A list specifying the layout of the combined plot. Default is `list(heights = c(1, 2))`, which sets the relative heights of the two plots.

annotation

A list specifying the annotation for the combined plot. Default is `list(tag_levels = "A")`, which adds alphabetic tags to the subplots.

Value

A list containing:

plot

A ggplot2 object representing the combined decision tree and policy relationship graph

Details

This function performs the following steps: 1. Generates policy results using `margot_plot_policy_tree` 2. Creates a decision tree plot using `margot_plot_decision_tree` 3. Combines the plots using the patchwork package 4. Applies the specified layout and annotation to the combined plot

See also

Examples

# Assuming mc_test and model_name are already defined:
result <- margot_plot_policy_combo(mc_test, "MyModel")
#> Error in eval(expr, envir, enclos): object 'mc_test' not found
print(result$plot)
#> Error in eval(expr, envir, enclos): object 'result' not found