Create Ordered Variable Based on Quantile Breaks or Custom Breaks with Informative Labels

This function divides a numeric variable into categories based on either quantile breaks or custom-specified breaks, and creates an ordered factor variable with informative labels.

Usage

create_ordered_variable(
  df,
  var_name,
  n_divisions = NULL,
  cutpoint_inclusive = "upper",
  ties.method = NULL,
  custom_breaks = NULL
)

Arguments

df

A data frame containing the variable to be divided into categories.

var_name

The name of the variable within the data frame to divide into categories.

n_divisions

The number of quantile divisions to create. Required if custom_breaks is not provided.

cutpoint_inclusive

A character string specifying whether cutpoints should be included in the lower or upper category. Must be either "lower" or "upper". Default is "upper".

ties.method

A character string specifying how ties should be handled when calculating quantiles. Must be one of "first", "last", "random", "ordered", or "average". If NULL (default), it will be set to "last" if cutpoint_inclusive is "upper", and "first" if cutpoint_inclusive is "lower".

custom_breaks

A numeric vector of break points to use for categorization. If provided, this overrides the quantile-based division specified by n_divisions.

Value

The input data frame with an additional column representing the ordered factor variable. The new column name will be the original variable name with "_cat" appended.

Details

The function creates categories based on either quantile breaks or custom-specified breaks. It then categorizes the values into these breaks, creating an ordered factor variable.

When using quantile breaks (custom_breaks is NULL): - The function creates quantile breaks based on the specified number of divisions. - If the number of unique values in the variable is less than the requested number of divisions, the function will adjust the number of divisions and issue a warning.

When using custom breaks: - If the lowest or highest values in the data are not included in the custom breaks, they will be automatically added and a warning will be issued. - Custom breaks are sorted and de-duplicated before use.

The `cutpoint_inclusive` parameter determines whether the cutpoint values themselves are included in the lower or upper category. For example, if a cutpoint is 40: - With "lower", the interval would be [40, 50) - With "upper", the interval would be (30, 40]

The `ties.method` parameter specifies how to handle tied values when calculating quantiles. It's recommended to use a method consistent with `cutpoint_inclusive`: - Use "first" or "last" to ensure consistency with "lower" or "upper" respectively. - "average" is not recommended as it may lead to inconsistent results.

Note

- Using "average" for `ties.method` will generate a warning due to potential inconsistencies. - When using custom breaks, ensure that the breaks cover the entire range of the data to avoid unexpected behavior.

Examples

df <- data.frame(x = rnorm(100))

# Using quantile breaks
result1 <- create_ordered_variable(df, "x", n_divisions = 4)
#> Summary of x_cat :
#> 
#> [-2.6,-0.4]  (-0.4,0.1]   (0.1,0.6]   (0.6,2.8] 
#>          25          25          25          25 

# Using custom breaks
result2 <- create_ordered_variable(df, "x", custom_breaks = c(-2, -1, 0, 1, 2))
#> Warning: Lowest break is higher than the minimum value. Adding minimum value to breaks.
#> Warning: Highest break is lower than the maximum value. Adding maximum value to breaks.
#> Summary of x_cat :
#> 
#> [-2.6,-2.0] (-2.0,-1.0]  (-1.0,0.0]   (0.0,1.0]   (1.0,2.0]   (2.0,2.8] 
#>           3          11          32          35          16           3 

# Explicitly setting cutpoint_inclusive and ties.method
result3 <- create_ordered_variable(df, "x", n_divisions = 3,
                                   cutpoint_inclusive = "lower",
                                   ties.method = "first")
#> Summary of x_cat :
#> 
#> [-2.6,-0.2)  [-0.2,0.5)   [0.5,2.8] 
#>          33          33          34