Run Multiple Generalized Random Forest (GRF) Causal Forest Models with Enhanced Features

This function is a wrapper for grf::causal_forest that runs multiple GRF causal forest models for specified outcome variables. It calculates average treatment effects, tests calibration, creates custom evaluation tables, and includes additional features such as tau.hat estimates, RATE calculations, policy trees, variable importance rankings, best linear projections, and Qini curves. It also prepares data for policy tree visualization, using a specified proportion of the data for training.

Usage

margot_causal_forest(
  data,
  outcome_vars,
  covariates,
  W,
  weights,
  grf_defaults = list(),
  save_data = FALSE,
  compute_rate = TRUE,
  top_n_vars = 10,
  save_models = FALSE,
  train_proportion = 0.8,
  propensity_threshold = 0.05
)

Arguments

data

A data frame containing all necessary variables.

outcome_vars

A character vector of outcome variable names to be modeled.

covariates

A matrix of covariates to be used in the GRF models.

W

A vector of binary treatment assignments.

weights

A vector of weights for the observations.

grf_defaults

A list of default parameters for the GRF models.

save_data

Logical indicating whether to save data, covariates, and weights. Default is FALSE.

compute_rate

Logical indicating whether to compute RATE for each model. Default is TRUE.

top_n_vars

Integer specifying the number of top variables to use for additional computations. Default is 10.

save_models

Logical indicating whether to save the full GRF model objects. Default is FALSE.

train_proportion

Numeric value between 0 and 1 indicating the proportion of non-missing data to use for training policy trees. Default is 0.8.

Value

A list containing:

results

A list of model results, one for each outcome variable. Each model result includes:

ateAverage treatment effect

test_calibrationCalibration test results

custom_tableCustom evaluation table

tau_hatIndividual treatment effect estimates

rate_resultRank Average Treatment Effect (if compute_rate is TRUE)

dr_scoresDouble robust scores

policy_tree_depth_1Policy tree of depth 1

top_varsTop variables by importance

blp_topBest linear projection results for top variables

policy_tree_depth_2Policy tree of depth 2, trained on train_proportion of non-missing data

plot_dataData prepared for policy tree visualization, using the remaining proportion of non-missing data

qini_dataData frame containing Qini curve data for plotting

combined_table

A data frame combining all custom evaluation tables.

outcome_vars

The character vector of outcome variable names that were modeled.

not_missing

A vector of indices for complete cases.

data

The input data (if save_data is TRUE).

covariates

The input covariates (if save_data is TRUE).

weights

The input weights (if save_data is TRUE).

full_models

A list of full GRF model objects (if save_models is TRUE).