Visualise Causal Effect Estimates with E-Values

Creates a graph of multiple causal effect estimates, organised by effect size. It is designed to work with the output of the `group_tab` function, allowing for visual comparison across estimates categorized as "positive", "negative", or "zero_crossing" based on their reliability and effect size. Supports both risk difference (RD) and risk ratio (RR) scales.

Usage

margot_plot(
  .data,
  type = c("RD", "RR"),
  title,
  subtitle,
  xlab,
  ylab,
  estimate_scale = 1,
  base_size = 11,
  text_size = 2.75,
  point_size = 0.5,
  title_size = 10,
  subtitle_size = 9,
  legend_text_size = 6,
  legend_title_size = 6,
  x_offset = ifelse(type == "RR", 0, -1.75),
  x_lim_lo = ifelse(type == "RR", 0.1, -1.75),
  x_lim_hi = ifelse(type == "RR", 2.5, 1)
)

Arguments

.data

Data frame output from `group_tab` function containing categorized causal effect estimates.

type

Character string indicating the scale of effect estimates to be visualised: "RD" for risk difference or "RR" for risk ratio. Default is "RD".

title

The main title for the plot.

subtitle

The subtitle for the plot.

xlab

Label for the x-axis, typically denoting the scale or measurement of effect estimates.

ylab

Label for the y-axis, generally left blank as outcomes are used as labels.

estimate_scale

Numeric value for adjusting the text offset of estimate labels. Default is 1.

base_size

Numeric value for the base font size in the plot. Default is 11.

text_size

Numeric value for the size of estimate labels. Default is 2.75.

point_size

Numeric value for the size of points representing the estimates. Default is 0.5.

title_size

Numeric value for the font size of the plot title. Default is 10.

subtitle_size

Numeric value for the font size of the plot subtitle. Default is 9.

legend_text_size

Numeric value for the font size of the legend text. Default is 6.

legend_title_size

Numeric value for the font size of the legend title. Default is 6.

x_offset

Numeric value for horizontal adjustment of estimate labels. Adjusts based on `type`.

x_lim_lo

Numeric value for the lower limit of the x-axis. Adjusts based on `type`.

x_lim_hi

Numeric value for the upper limit of the x-axis. Adjusts based on `type`.

Value

A ggplot object visualising the causal effect estimates with error bars and categorization labels, ready for further modification or saving.

Examples

 if (FALSE) {
title <- "Religious Service At Least Once Per Week vs None"

# Plotting perceived social support with Risk Difference (RD) scale
plot_group_tab_all_perceived_support <- margot_plot(
  group_tab_all_perceived_support,
  type = "RD",
  title = title,
  subtitle = "Perceived Social Support",
  xlab = "",
  ylab = "",
  estimate_scale = 1,
  base_size = 18,
  text_size = 4.5,
  point_size = 3.5,
  title_size = 20,
  subtitle_size = 16,
  legend_text_size = 10,
  legend_title_size = 10,
  x_offset = -.5,
  x_lim_lo = -.5,
  x_lim_hi =  .5
)

# Plot for support received from others with Risk Ratio (RR) scale
plot_group_tab_all_received_time <- margot_plot(
  group_tab_contrast_received_time,
  type = "RR",
  title = title,
  subtitle = "Support Received From Others: Time",
  xlab = "",
  ylab = "",
  estimate_scale = 1,
  base_size = 18,
  text_size = 4.5,
  point_size = 3.5,
  title_size = 20,
  subtitle_size = 16,
  legend_text_size = 10,
  legend_title_size = 10,
  x_offset = 0,
  x_lim_lo = 0,
  x_lim_hi =  4
)

print(plot_group_tab_all_perceived_support)
print(plot_group_tab_all_received_time)
}