Plot Policy Tree Results

This function creates a visualization of policy tree results from a causal forest or multi-arm causal forest model. It generates two plots showing the relationships between the top three split variables, with points colored by the predicted optimal treatment.

Usage

margot_plot_policy_tree(
  mc_test,
  model_name,
  color_scale = ggokabeito::scale_colour_okabe_ito(),
  point_alpha = 0.6,
  theme_function = theme_classic,
  title_size = 14,
  subtitle_size = 12,
  axis_title_size = 10,
  legend_title_size = 10,
  jitter_width = 0.3,
  jitter_height = 0.3
)

Arguments

mc_test

A list containing the results from a multi-arm causal forest model.

model_name

A string specifying which model's results to plot.

color_scale

A ggplot2 color scale. Default is ggokabeito::scale_colour_okabe_ito().

point_alpha

Numeric value between 0 and 1 for point transparency. Default is 0.6.

theme_function

A ggplot2 theme function. Default is theme_classic.

title_size

Numeric value for the size of the plot title. Default is 14.

subtitle_size

Numeric value for the size of the plot subtitle. Default is 12.

axis_title_size

Numeric value for the size of axis titles. Default is 10.

legend_title_size

Numeric value for the size of the legend title. Default is 10.

jitter_width

Numeric value for the amount of horizontal jitter. Default is 0.3.

jitter_height

Numeric value for the amount of vertical jitter. Default is 0.3.

Value

A ggplot object containing two side-by-side plots of the policy tree results.

Examples

plot <- plot_policy_tree_results(mc_test, "model_t2_belong_z")
#> Error in plot_policy_tree_results(mc_test, "model_t2_belong_z"): could not find function "plot_policy_tree_results"
print(plot)
#> function (x, y, ...) 
#> UseMethod("plot")
#> <bytecode: 0x1057ec1c0>
#> <environment: namespace:base>