Format RATE Results into Readable Tables

Format RATE Results into Readable Tables

Usage

margot_rate(
  models,
  label_mapping = NULL,
  remove_tx_prefix = TRUE,
  remove_z_suffix = TRUE,
  use_title_case = TRUE,
  remove_underscores = TRUE,
  round_digits = 3,
  highlight_significant = FALSE
)

Arguments

models

A list containing model results from which to extract RATE information.

label_mapping

An optional named vector for mapping variable names to readable labels.

remove_tx_prefix

Logical indicating whether to remove ‘tx_’ prefix from outcome names. Default is TRUE.

remove_z_suffix

Logical indicating whether to remove ‘_z’ suffix from outcome names. Default is TRUE.

use_title_case

Logical indicating whether to convert outcome names to title case. Default is TRUE.

remove_underscores

Logical indicating whether to replace underscores with spaces. Default is TRUE.

round_digits

Integer specifying the number of decimal places for rounding. Default is 3.

highlight_significant

Logical indicating whether to add asterisks to outcomes with confidence intervals that don‘t’ cross zero. Default is FALSE.

Value

A list containing two tables: rate_result and rate_qini.